--- projectextractor/src/main/java/com/minerprojects/badsmelldetector/ExecutionConfig.java
@@ -1,9 +1,0 @@
-package com.minerprojects.badsmelldetector;
-
-public class ExecutionConfig {
-
-    public static String PROJECT_PATH = "C:\\Users\\kleit\\OneDrive\\Documentos\\tcc\\java-paser-refactoring-and-comments\\projectextractor\\tmp\\";
-
-    public static final String PMD_PATH = "C:\\pmd-bin-7.8.0\\bin";
-
-}
