+++ projectextractor/src/main/java/com/minerprojects/pmddetector/ExecutionConfig.java
@@ -1,0 +1,9 @@
+package com.minerprojects.pmddetector;
+
+public class ExecutionConfig {
+
+    public static String PROJECT_PATH = "C:\\Users\\kleit\\OneDrive\\Documentos\\tcc\\java-paser-refactoring-and-comments\\projectextractor\\tmp\\";
+
+    public static final String PMD_PATH = "C:\\pmd-bin-7.8.0\\bin";
+
+}
