+++ projectextractor/src/main/java/com/minerprojects/pmddetector/pmd/CyclomaticComplexity.java
@@ -1,0 +1,55 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.minerprojects.pmddetector.pmd;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+import org.springframework.web.client.RestTemplate;
+
+import com.minerprojects.CommitReporter;
+import com.minerprojects.PMDReporter;
+import com.minerprojects.data.DataPMD;
+import com.minerprojects.pmddetector.violation.ViolationPMD2;
+
+public class CyclomaticComplexity extends BadSmellPMD2 {
+
+    private static Logger logger = Logger.getLogger(CyclomaticComplexity.class.getName());
+
+    @Override
+    public String toString() {
+        return "CyclomaticComplexity{" + super.toString() + '}';
+    }
+
+    public CyclomaticComplexity() {
+    }
+
+    public CyclomaticComplexity(ViolationPMD2 violationPMD2) {
+        super(violationPMD2);
+    }
+
+    public static List<CyclomaticComplexity> extractCyclomaticComplexity(String projectDirectory,
+            String projectName, CommitReporter commit) {
+
+        String dir = projectDirectory.substring(0, projectDirectory.lastIndexOf(File.separator));
+        List<CyclomaticComplexity> result = new ArrayList();
+
+        try {
+
+            PMDReporter.analyzeFile(dir, projectName, "CyclomaticComplexity").forEach(violation -> {
+                PMDReporter.save(violation, projectName, commit);
+            });
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return result;
+    }
+}
