+++ projectextractor/src/main/java/com/minerprojects/pmddetector/violation/Violation.java
@@ -1,0 +1,56 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.minerprojects.pmddetector.violation;
+
+public class Violation {
+    private int beginLine;
+    private int endLine;
+    private int beginColumn;
+    private int endColumn;
+
+    public Violation() {
+    }
+
+    public Violation(int beginLine, int endLine, int beginColumn, int endColumn) {
+        this.beginLine = beginLine;
+        this.endLine = endLine;
+        this.beginColumn = beginColumn;
+        this.endColumn = endColumn;
+    }
+
+    public int getBeginLine() {
+        return beginLine;
+    }
+
+    public void setBeginLine(int beginLine) {
+        this.beginLine = beginLine;
+    }
+
+    public int getEndLine() {
+        return endLine;
+    }
+
+    public void setEndLine(int endLine) {
+        this.endLine = endLine;
+    }
+
+    public int getBeginColumn() {
+        return beginColumn;
+    }
+
+    public void setBeginColumn(int beginColumn) {
+        this.beginColumn = beginColumn;
+    }
+
+    public int getEndColumn() {
+        return endColumn;
+    }
+
+    public void setEndColumn(int endColumn) {
+        this.endColumn = endColumn;
+    }
+
+}
